Ok, this one does 101.8 kkeys/s on my DX4/100 instead of 97 kkeys/s for the previous 386/486 code.
In this core I modify the code itself at run-time to store
intermediate results instead of using the stack. Thus I can avoid
reading from memory, and it save some cycles.

This core will never be multithread-safe. That's not a big deal,
because there is very few 486 SMP boards anyway.

This speed-up is absolutely dependent on the correct alignment of
instructions to be modified. Thus I had to insert some nop's, and to
change some instructions to inflate by a few bytes the code.

It require some support from the compiler tools too. On most platforms, a
given piece of code isn't allowed to write in the .text section. On
linux, passing "-Xlinker -omagic" (!) to GCC at link time does the
trick. On Win32 there's a command-line utility in MS VC++ 5.0 which
can modify flags in the EXE header, but I don't recall the exact
command right now.
Without this support from the compiler / linker, your binary will
probably segfault as soon as it enters this code.