Change threshold calculations:   New option: fetch-time-threshold: will attempt to get enough work units     to keep the client busy for that many hours.   New option: fetch-workunit-threshold: will fetch specified number of     work units (rather than packets) for the client.
If the client hasn't had the chance to complete any blocks and a time
based threshold is being used, it will fall back first to a specified
work unit threshold, then to a "minimalistic" fallback (one work unit
per CPU for fetch; MAXBLOCKSPERBUFFER for flush).  If the client decides
that it is definitely going to fetch/flush packets now (including the
-fetch option), it will calculate the threshold by running a microbenchmark.

Any regularly completed block, benchmark, or core-selection microbenchmark
will be used to determine the ideal crunching speed of the client.  The
best result (for this run of the client) is used.

Time (in hours) is capped at two weeks, but in reality, after the time
threshold is used to calculate a workunit threshold, that work unit
threshold is capped at MAXBLOCKSPERBUFFER.

Removed option: fetch-threshold (is migrated into fetch-workunit-threshold)

Changed messages to reflect that work-units are the fundamental measure
of progress, rather than network packets.

Don't use client->inthreshold[] or client->outthreshold[] anymore.  Use
ClientGetInThreshold(client, contest) or ClientGetOutThreshold() instead.

Made the "microbench" in the core select routines go through the benchmark
code (so it will be picked up the threshold calculations).

Because of my ignorance for how to compute the "difficulty" of an unprocessed
OGR packet, time thresholds for OGR are disabled, and 1 work unit currently
equals 1 network packet for OGR.

Added a display of the approx. amount of time it will take to exhaust
the input buffer when displaying how many work units remain in the input
buffer.

Hopefully, this all works on everyone's machine.  ;-)