print cause of [OGR-] LoadState() failure.
IMO, failing LoadState() [which always causes the work to be discarded]
due to data errors is silly. Since the proxy network doesn't know that
the block was discarded due to bad data (all it knows is that it wasn't
returned), it will eventually re-issue it, and that in turn implies that
the contest will never close because some data was never (and could never
be) returned.

Since it is apparently impossible for keyservers to filter these bad-stub
errors (_E_FORMAT and _E_STUB) in the first place, what ogr_create()
really _should_ do is delegate these errors (and perhaps all errors) to
some later stage, which in turn would "truncate" the stub (sets total
number of nodes done to zero) and return RESULT_NOTHING.

(RetrieveState() could then optionally print an appropriate error message)